backend/requirements.txt
torch
torchvision
torchaudio
opencv-python
pillow
numpy
fastapi
python-multipart
uvicorn
requests
tqdm
